requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
PyPDF2
google-generativeai
python-dotenv
